Improving javadoc documentation for Rapids; several misplaced methods were relocated to where they would make actual sense
