+++ h2o-core/src/main/java/water/rapids/Env.java
@@ -2,14 +2,27 @@ package water.rapids;
 
 import water.*;
 import water.fvec.Frame;
-import water.rapids.ast.AstFunction;
-import water.rapids.ast.AstPrimitive;
-import water.rapids.ast.AstRoot;
+import water.rapids.ast.*;
+import water.rapids.ast.params.AstConst;
+import water.rapids.ast.prims.advmath.*;
+import water.rapids.ast.prims.assign.*;
+import water.rapids.ast.prims.math.*;
+import water.rapids.ast.prims.matrix.*;
+import water.rapids.ast.prims.misc.*;
+import water.rapids.ast.prims.mungers.*;
+import water.rapids.ast.prims.operators.*;
+import water.rapids.ast.prims.reducers.*;
+import water.rapids.ast.prims.repeaters.*;
+import water.rapids.ast.prims.search.*;
+import water.rapids.ast.prims.string.*;
+import water.rapids.ast.prims.time.*;
+import water.rapids.ast.prims.timeseries.*;
 import water.rapids.vals.ValFrame;
 import water.rapids.vals.ValFun;
 
 import java.io.Closeable;
 import java.util.ArrayList;
+import java.util.HashMap;
 
 /**
  * Execute a set of instructions in the context of an H2O cloud.
@@ -31,17 +44,241 @@ import java.util.ArrayList;
  */
 public class Env extends Iced {
 
-  // Session holds the ref-cnts across multiple executions.
+  // Session holds the ref-counts across multiple executions.
   public final Session _ses;
 
-  public Env(Session ses) {
-    _ses = ses;
-  }
+  // Current lexical scope lookup
+  public AstFunction _scope;
+
 
   // Frames that are alive in mid-execution; usually because we have evaluated
   // some first expression and need to hang onto it while evaluating the next
   // expression.
   private ArrayList<Frame> _stk = new ArrayList<>();
+
+  // Built-in constants, checked before other namespace lookups happen
+  private static final HashMap<String, AstPrimitive> PRIMS = new HashMap<>();
+  // Built-in primitives, done after other namespace lookups happen
+  private static final HashMap<String, AstParameter> CONSTS = new HashMap<>();
+
+  static void init(AstPrimitive ast) {
+    PRIMS.put(ast.str(), ast);
+  }
+
+  static {
+    // Constants
+    CONSTS.put("FALSE", AstConst.FALSE);
+    CONSTS.put("False", AstConst.FALSE);
+    CONSTS.put("false", AstConst.FALSE);
+    CONSTS.put("TRUE", AstConst.TRUE);
+    CONSTS.put("True", AstConst.TRUE);
+    CONSTS.put("true", AstConst.TRUE);
+    CONSTS.put("NaN", AstConst.NAN);
+    CONSTS.put("NA", AstConst.NAN);
+    CONSTS.put("PI", AstConst.PI);
+    CONSTS.put("Pi", AstConst.PI);
+
+    // Standard math functions
+    init(new AstAbs());
+    init(new AstAcos());
+    init(new AstAcosh());
+    init(new AstAsin());
+    init(new AstAsinh());
+    init(new AstAtan());
+    init(new AstAtanh());
+    init(new AstCeiling());
+    init(new AstCos());
+    init(new AstCosh());
+    init(new AstCosPi());
+    init(new AstDiGamma());
+    init(new AstExp());
+    init(new AstExpm1());
+    init(new AstFloor());
+    init(new AstGamma());
+    init(new AstLGamma());
+    init(new AstLog());
+    init(new AstLog1P());
+    init(new AstLog2());
+    init(new AstLog10());
+    init(new AstNoOp());
+    init(new AstNot());
+    init(new AstRound());
+    init(new AstSgn());
+    init(new AstSignif());
+    init(new AstSin());
+    init(new AstSinh());
+    init(new AstSinPi());
+    init(new AstSqrt());
+    init(new AstTan());
+    init(new AstTanh());
+    init(new AstTanPi());
+    init(new AstTriGamma());
+    init(new AstTrunc());
+
+    // Math binary operators
+    init(new AstAnd());
+    init(new AstDiv());
+    init(new AstEq());
+    init(new AstGe());
+    init(new AstGt());
+    init(new AstIntDiv());
+    init(new AstIntDivR());
+    init(new AstLAnd());
+    init(new AstLe());
+    init(new AstLOr());
+    init(new AstLt());
+    init(new AstMod());
+    init(new AstModR());
+    init(new AstMul());
+    init(new AstNe());
+    init(new AstOr());
+    init(new AstPlus());
+    init(new AstPow());
+    init(new AstSub());
+    init(new AstIfElse());
+    init(new AstIfElse()); // this one is ternary
+
+    // Reducers
+    init(new AstAll());
+    init(new AstAny());
+    init(new AstAnyNa());
+    init(new AstCumMax());
+    init(new AstCumMin());
+    init(new AstCumProd());
+    init(new AstCumSum());
+    init(new AstMad());
+    init(new AstMax());
+    init(new AstMaxNa());
+    init(new AstMean());
+    init(new AstMedian());
+    init(new AstMin());
+    init(new AstMinNa());
+    init(new AstNaCnt());
+    init(new AstProd());
+    init(new AstProdNa());
+    init(new AstSdev());
+    init(new AstSum());
+    init(new AstSumNa());
+
+    // Time
+    init(new AstAsDate());
+    init(new AstDay());
+    init(new AstDayOfWeek());
+    init(new AstGetTimeZone());
+    init(new AstHour());
+    init(new AstListTimeZones());
+    init(new AstMillis());
+    init(new AstMinute());
+    init(new AstMktime());
+    init(new AstMonth());
+    init(new AstSecond());
+    init(new AstSetTimeZone());
+    init(new AstWeek());
+    init(new AstYear());
+
+    // Time Series
+    init(new AstDiffLag1());
+
+    // Advanced Math
+    init(new AstCorrelation());
+    init(new AstHist());
+    init(new AstImpute());
+    init(new AstKFold());
+    init(new AstMode());
+    init(new AstSkewness());
+    init(new AstKurtosis());
+    init(new AstModuloKFold());
+    init(new AstQtile());
+    init(new AstRunif());
+    init(new AstSort());
+    init(new AstStratifiedKFold());
+    init(new AstStratifiedSplit());
+    init(new AstTable());
+    init(new AstUnique());
+    init(new AstVariance());
+
+    // Generic data mungers
+    init(new AstAnyFactor());
+    init(new AstApply());
+    init(new AstAsFactor());
+    init(new AstAsCharacter());
+    init(new AstAsNumeric());
+    init(new AstCBind());
+    init(new AstColNames());
+    init(new AstColPySlice());
+    init(new AstColSlice());
+    init(new AstCut());
+    init(new AstDdply());
+    init(new AstFilterNaCols());
+    init(new AstFlatten());
+    init(new AstGetrow());
+    init(new AstGroup());
+    init(new AstGroupedPermute());
+    init(new AstIsCharacter());
+    init(new AstIsFactor());
+    init(new AstIsNa());
+    init(new AstIsNumeric());
+    init(new AstLevels());
+    init(new AstMerge());
+    init(new AstNaOmit());
+    init(new AstColumnsByType());
+    init(new AstNcol());
+    init(new AstNLevels());
+    init(new AstNrow());
+    init(new AstRBind());
+    init(new AstReLevel());
+    init(new AstRename());
+    init(new AstRowSlice());
+    init(new AstScale());
+    init(new AstSetDomain());
+    init(new AstSetLevel());
+
+    // Assignment; all of these lean heavily on Copy-On-Write optimizations.
+    init(new AstAppend());      // Add a column
+    init(new AstAssign());      // Overwrite a global
+    init(new AstRectangleAssign()); // Overwrite a rectangular slice
+    init(new AstRm());          // Remove a frame, but maintain internal sharing
+    init(new AstTmpAssign());   // Create a new immutable tmp frame
+
+    // Matrix functions
+    init(new AstTranspose());
+    init(new AstMMult());
+
+    // String functions
+    init(new AstCountMatches());
+    init(new AstCountSubstringsWords());
+    init(new AstEntropy());
+    init(new AstLStrip());
+    init(new AstReplaceAll());
+    init(new AstReplaceFirst());
+    init(new AstRStrip());
+    init(new AstStrLength());
+    init(new AstStrSplit());
+    init(new AstSubstring());
+    init(new AstToLower());
+    init(new AstToUpper());
+    init(new AstTrim());
+
+    // Miscellaneous
+    init(new AstComma());
+    init(new AstLs());
+
+    // Search
+    init(new AstMatch());
+    init(new AstWhich());
+
+    // Repeaters
+    init(new AstRepLen());
+    init(new AstSeq());
+    init(new AstSeqLen());
+
+  }
+
+
+
+  public Env(Session ses) {
+    _ses = ses;
+  }
 
   public int sp() {
     return _stk.size();
@@ -112,16 +349,14 @@ public class Env extends Iced {
   // ----
   // Variable lookup
 
-  public AstFunction _scope;                // Current lexical scope lookup
-
   public Val lookup(String id) {
     // Lexically scoped functions first
     Val val = _scope == null ? null : _scope.lookup(id);
     if (val != null) return val;
 
     // disallow TRUE/FALSE/NA to be overwritten by keys in the DKV... just way way saner this way
-    if (AstRoot.CONSTS.containsKey(id)) {
-      return AstRoot.CONSTS.get(id).exec(this);
+    if (CONSTS.containsKey(id)) {
+      return CONSTS.get(id).exec(this);
     }
 
     // Now the DKV
@@ -134,7 +369,7 @@ public class Env extends Iced {
     }
 
     // Now the built-ins
-    AstPrimitive ast = AstRoot.PRIMS.get(id);
+    AstPrimitive ast = PRIMS.get(id);
     if (ast != null)
       return new ValFun(ast);
 
@@ -159,23 +394,4 @@ public class Env extends Iced {
     return s + "}";
   }
 
-  public AutoBuffer write_impl(AutoBuffer ab) {
-    //ab.put4(_globals.size());   for( Vec vec : _globals ) ab.put(vec._key);
-    //ab.put4(sp());              for( Frame fr : _stk ) ab.put(fr);
-    //return ab;
-    throw water.H2O.unimpl();
-  }
-
-  public Env read_impl(AutoBuffer ab) {
-    //_globals = new HashSet<>();
-    //int len=ab.get4();
-    //for( int i=0; i<len; i++ )
-    //  _globals.add((Vec)DKV.getGet((Key)ab.get()));
-    //_stk = new ArrayList<>();
-    //len=ab.get4();
-    //for( int i=0; i<len; i++ )
-    //  _stk.add((Frame)ab.get());
-    //return this;
-    throw water.H2O.unimpl();
-  }
 }
