+++ h2o-core/src/main/java/water/rapids/SplitByMSBLocal.java
@@ -1,5 +1,11 @@
 package water.rapids;
 
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+//
+// TODO: move to water.rapids.ast.prims.mungers (this is a helper class for AstMerge / AstSort)
+//
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
 import water.*;
 import water.fvec.Chunk;
 import water.util.ArrayUtils;
