+++ h2o-core/src/main/java/water/rapids/Val.java
@@ -2,6 +2,7 @@ package water.rapids;
 
 import water.fvec.Frame;
 import water.Iced;
+import water.rapids.ast.AstPrimitive;
 import water.rapids.ast.AstRoot;
 
 /**
@@ -35,7 +36,7 @@ abstract public class Val extends Iced {
   public String[] getStrs()  { throw badValue("String array"); }
   public Frame    getFrame() { throw badValue("Frame"); }
   public double[] getRow()   { throw badValue("Row"); }
-  public AstRoot  getFun()   { throw badValue("function"); }
+  public AstPrimitive getFun() { throw badValue("function"); }
 
   private IllegalArgumentException badValue(String expectedType) {
     return new IllegalArgumentException("Expected a " + expectedType + " but found a " + getClass());
