+++ h2o-core/src/main/java/water/rapids/ast/AstFrame.java
@@ -39,8 +39,4 @@ public class AstFrame extends AstRoot {
     return env.returning(_fr);
   }
 
-  @Override
-  public int nargs() {
-    return 1;
-  }
 }
