+++ h2o-core/src/main/java/water/rapids/ast/AstFunction.java
@@ -12,6 +12,8 @@ import java.util.ArrayList;
  * Define a function
  * Syntax: { ids... . expr }
  * IDs are bound within expr
+ *
+ * TODO: rename to AstUserDefinedFunction
  */
 public class AstFunction extends AstPrimitive {
   final String[] _ids;          // Identifier names
