+++ h2o-core/src/main/java/water/rapids/ast/params/AstId.java
@@ -1,5 +1,6 @@
 package water.rapids.ast.params;
 
+import water.fvec.Frame;
 import water.rapids.Env;
 import water.rapids.Rapids;
 import water.rapids.Val;
@@ -23,6 +24,10 @@ public class AstId extends AstParameter {
     _id = id;
   }
 
+  public AstId(Frame f) {
+    _id = f._key.toString();
+  }
+
   @Override
   public String str() {
     return _id;
@@ -31,11 +36,6 @@ public class AstId extends AstParameter {
   @Override
   public Val exec(Env env) {
     return env.returning(env.lookup(_id));
-  }
-
-  @Override
-  public int nargs() {
-    return 1;
   }
 
   @Override
