+++ h2o-core/src/main/java/water/rapids/ast/prims/mungers/AstColPySlice.java
@@ -3,6 +3,7 @@ package water.rapids.ast.prims.mungers;
 import water.fvec.Frame;
 import water.rapids.Env;
 import water.rapids.Val;
+import water.rapids.ast.AstParameter;
 import water.rapids.ast.AstRoot;
 import water.rapids.vals.ValFrame;
 import water.rapids.vals.ValRow;
@@ -34,12 +35,13 @@ public class AstColPySlice extends AstPrimitive {
   @Override
   public Val apply(Env env, Env.StackHelp stk, AstRoot asts[]) {
     Val v = stk.track(asts[1].exec(env));
+    AstParameter col_list = (AstParameter) asts[2];
     if (v instanceof ValRow) {
       ValRow vv = (ValRow) v;
-      return vv.slice(asts[2].columns(vv.getNames()));
+      return vv.slice(col_list.columns(vv.getNames()));
     }
     Frame fr = v.getFrame();
-    int[] cols = asts[2].columns(fr.names());
+    int[] cols = col_list.columns(fr.names());
 
     Frame fr2 = new Frame();
     if (cols.length == 0)        // Empty inclusion list?
