+++ h2o-core/src/main/java/water/rapids/ast/prims/mungers/AstSort.java
@@ -4,6 +4,7 @@ import water.fvec.*;
 import water.rapids.Env;
 import water.rapids.Merge;
 import water.rapids.Val;
+import water.rapids.ast.AstParameter;
 import water.rapids.ast.AstPrimitive;
 import water.rapids.ast.AstRoot;
 import water.rapids.vals.ValFrame;
@@ -18,7 +19,7 @@ public class AstSort extends AstPrimitive {
 
   @Override public ValFrame apply(Env env, Env.StackHelp stk, AstRoot asts[]) {
     Frame fr = stk.track(asts[1].exec(env)).getFrame();
-    int[] cols = asts[2].columns(fr.names());
+    int[] cols = ((AstParameter)asts[2]).columns(fr.names());
     return new ValFrame(Merge.sort(fr,cols));
   }
 }
